+++ native/storybook/storybook.js
@@ -90,6 +90,7 @@ configure(() => {
   require('../packages/react-native-bpk-component-text/stories');
   require('../packages/react-native-bpk-theming/stories');
   require('../packages/react-native-bpk-component-touchable-overlay/stories');
+  require('../packages/react-native-bpk-component-touchable-native-feedback/stories');
 }, module);
 /* eslint-enable global-require */